feat :: 실행되지 않는 알람 예외 적용

+++ src/main/java/team/jeonghokim/daedongyeojido/domain/club/service/DeleteClubAlarmService.java
@@ -4,9 +4,11 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team.jeonghokim.daedongyeojido.domain.alarm.domain.ClubAlarm;
+import team.jeonghokim.daedongyeojido.domain.alarm.domain.enums.AlarmCategory;
 import team.jeonghokim.daedongyeojido.domain.alarm.domain.repository.ClubAlarmRepository;
 import team.jeonghokim.daedongyeojido.domain.alarm.exception.AlarmAccessDeniedException;
 import team.jeonghokim.daedongyeojido.domain.alarm.exception.AlarmNotFoundException;
+import team.jeonghokim.daedongyeojido.domain.alarm.exception.CannotDeleteAlarmException;
 import team.jeonghokim.daedongyeojido.domain.club.exception.UserNotInClubException;
 import team.jeonghokim.daedongyeojido.domain.user.domain.User;
 import team.jeonghokim.daedongyeojido.domain.user.facade.UserFacade;
@@ -31,6 +33,10 @@ public class DeleteClubAlarmService {
 
         if (!clubAlarm.getClub().getId().equals(user.getClub().getId())) {
             throw AlarmAccessDeniedException.EXCEPTION;
+        }
+
+        if (!clubAlarm.getAlarmCategory().equals(AlarmCategory.COMMON) && !clubAlarm.isExecuted()) {
+            throw CannotDeleteAlarmException.EXCEPTION;
         }
 
         clubAlarmRepository.delete(clubAlarm);
